Adds an extra menu name and menu description to locations
Resolves #23

Name shows up when you are at the location, menu_name is used in the
menu, and menu_description is placed in the description box when the
menu item is highlighted. If they are left out, it defaults to name.

Dungeons now fill in menu_name and menu_description to make things more
clear.