Changes the behavior when a story asks for activity, but activity is off
Resolves #88
Now, if activity is off either because the watch doesn’t support it or
the user turns it off, we treat the player as active instead of
inactive. This allows the playing of stories by those who don’t want or
can’t be active during their play session. It also makes testing in the
emulator easier.